+++ packages/moonstone/CHANGELOG.md
@@ -27,6 +27,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/VirtualList` to mute its container instead of disabling it during scroll events
 - `moonstone/VirtualList`, `moonstone/VirtualGridList`, and `moonstone/Scroller` to continue scrolling when holding down the paging controls
 - `moonstone/VirtualList` to require a `component` prop and not have a default value
+- `moonstone/Picker` to continuously change when a button is held down by adding `ui/Holdable`.
 
 ### Fixed
 
@@ -34,6 +35,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - Bug where a synchronized marquee whose content fit the available space would prevent restarting of the marquees
 - `moonstone/Input` to show an ellipsis on the correct side based on the text directionality of the `value` or `placeholder` content.
 - `moonstone/VirtualList` and `moonstone/VirtualGridList` to prevent unwanted scrolling when focused with the pointer
+- `moonstone/Picker` to remove fingernail when a the pointer is held down, but the pointer is moved off the `joined` picker.
 
 ## [1.0.0-alpha.5] - 2016-12-16
 
@@ -97,7 +99,6 @@ imported, as they already import the general mixins.
 - `moonstone/ToggleItem`'s `inline` version to have a `max-width` of `240px`
 - `moonstone/Input` to use `<div>` instead of `<label>` for wrapping components. No change to
 	functionality, only markup.
-- `moonstone/Picker`'s `moonstone/IconButton` now has a `onMouseLeave` prop to remove the pressed state.  
 
 ### Removed
 
@@ -124,4 +125,3 @@ This version includes a lot of refactoring from the previous release. Developers
 
 - Many components were fixed, polished, updated and documented
 - Inline docs updated to be more consistent and comprehensive
-- Fixed `Picker` pressed state staying active when `onMouseUp` occurs outside of button.